web e fix de desktop(sidebar para o mecanico e rececionista)

+++ src/main/java/com/example/projeto2/Repo/ReparacaoRepository.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface ReparacaoRepository extends JpaRepository<Reparacao, BigDecimal> {
+public interface ReparacaoRepository extends JpaRepository<Reparacao, Long> {
     int countByDataInicioLessThanEqualAndDataFimGreaterThanEqual(LocalDate date, LocalDate endDate);
     int countByDataFimLessThanEqualAndEstado(LocalDate date, String estado);
     List<Reparacao> findByDataInicioLessThanEqualAndDataFimGreaterThanEqual(LocalDate date, LocalDate endDate);
